+++ src/app/(auth)/app/(sidebar)/projects/configuration/accessories/_components/table.tsx
@@ -21,9 +21,8 @@ interface TableProps {
 
 const Table = ({ onEditClick, onDeleteClick, onAddClick }: TableProps) => {
   const router = useRouter();
-  const searchParams = useSearchParams();
-  const offset = Number(searchParams.get('offset') || 0);
   const [search, setSearch] = useQueryParam('search');
+
 
   const fetcher = async ({ offset, limit }: { offset: number; limit: number }) => {
     const res = await getAccessories({ offset, limit, search: search || undefined });
@@ -167,7 +166,7 @@ const Table = ({ onEditClick, onDeleteClick, onAddClick }: TableProps) => {
         </Button>
       </div>
       <TableData<Accessory>
-        queryKey={['accessories', search, offset]}
+        queryKey={['accessories', search]}
         fetcher={fetcher}
         columns={columns}
         renderCard={renderCard}